ENH: Deformation field takes Float images as input JIRA #OTB-555

+++ Applications/DisparityMap/otbStereoFramework.cxx
@@ -87,7 +87,7 @@ public:
   typedef otb::StereorectificationDeformationFieldSource
     <FloatImageType,FloatVectorImageType>     DeformationFieldSourceType;
 
-  typedef itk::Vector<double,2>               DeformationType;
+  typedef itk::Vector<float,2>               DeformationType;
   typedef otb::Image<DeformationType>         DeformationFieldType;
 
   typedef itk::VectorCastImageFilter
